+++ dbus-serialbattery/qml/gui-v2/3.5x/PageBattery.qml
@@ -116,10 +116,25 @@ Page {
 			ListQuantityGroup {
 				text: CommonWords.battery
 				textModel: [
-					{ value: root.battery.voltage, unit: VenusOS.Units_Volt_DC },
-					{ value: root.battery.current, unit: VenusOS.Units_Amp },
-					{ value: root.battery.power, unit: VenusOS.Units_Watt }
+					{ value: batteryVoltage.value, unit: VenusOS.Units_Volt_DC },
+					{ value: batteryCurrent.value, unit: VenusOS.Units_Amp },
+					{ value: batteryPower.value, unit: VenusOS.Units_Watt }
 				]
+
+				VeQuickItem {
+					id: batteryVoltage
+					uid: root.battery.serviceUid + "/Dc/0/Voltage"
+				}
+
+				VeQuickItem {
+					id: batteryCurrent
+					uid: root.battery.serviceUid + "/Dc/0/Current"
+				}
+
+				VeQuickItem {
+					id: batteryPower
+					uid: root.battery.serviceUid + "/Dc/0/Power"
+				}
 			}
 
 			ListQuantityItem {
@@ -132,14 +147,19 @@ Page {
 
 			ListQuantityItem {
 				readonly property VeQuickItem _n2kDeviceInstance: VeQuickItem {
-					uid: battery.serviceUid + "/N2kDeviceInstance"
+					uid: root.battery.serviceUid + "/N2kDeviceInstance"
 				}
 
 				//% "System voltage"
 				text: qsTrId("devicelist_battery_system_voltage")
 				dataItem.uid: BackendConnection.serviceUidFromName("com.victronenergy.battery.lynxparallel" + _n2kDeviceInstance.value, _n2kDeviceInstance.value) + "/Dc/0/Voltage"
-				allowed: defaultAllowed && !root.isParallelBms && root.battery.state === VenusOS.Battery_State_Pending
+				allowed: defaultAllowed && !root.isParallelBms && batteryState.value === VenusOS.Battery_State_Pending
 				unit: VenusOS.Units_Volt_DC
+
+				VeQuickItem {
+					id: batteryState
+					uid: root.battery.serviceUid + "/State"
+				}
 			}
 
 			ListTextItem {
@@ -152,7 +172,7 @@ Page {
 
 			ListQuantityItem {
 				text: CommonWords.state_of_charge
-				value: root.battery.stateOfCharge
+				dataItem.uid: root.battery.serviceUid + "/Soc"
 				unit: VenusOS.Units_Percentage
 			}
 
@@ -167,13 +187,6 @@ Page {
 			ListTemperatureItem {
 				text: CommonWords.battery_temperature
 				dataItem.uid: root.battery.serviceUid + "/Dc/0/Temperature"
-				allowed: defaultAllowed && dataItem.isValid
-				unit: Global.systemSettings.temperatureUnit
-			}
-
-			ListTemperatureItem {
-				text: "MOSFET Temperature"
-				dataItem.uid: root.battery.serviceUid + "/System/MOSTemperature"
 				allowed: defaultAllowed && dataItem.isValid
 				unit: Global.systemSettings.temperatureUnit
 			}
@@ -205,7 +218,7 @@ Page {
 				//% "Top section voltage"
 				text: qsTrId("battery_top_section_voltage")
 				allowed: midVoltage.isValid
-				value: midVoltage.isValid && !isNaN(root.battery.voltage) ? root.battery.voltage - midVoltage.value : NaN
+				value: midVoltage.isValid && batteryVoltage.isValid ? batteryVoltage.value - midVoltage.value : NaN
 				unit: VenusOS.Units_Volt_DC
 			}
 
@@ -246,55 +259,7 @@ Page {
 				text: qsTrId("battery_time_to_go")
 				dataItem.uid: root.battery.serviceUid + "/TimeToGo"
 				allowed: defaultAllowed && dataItem.seen
-				secondaryText: Utils.secondsToString(root.battery.timeToGo)
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 0%"
-				text: "Time-to-SoC 0%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/0"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 10%"
-				text: "Time-to-SoC 10%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/10"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 20%"
-				text: "Time-to-SoC 20%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/20"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 80%"
-				text: "Time-to-SoC 80%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/80"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 90%"
-				text: "Time-to-SoC 90%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/90"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
-			}
-
-			ListTextItem {
-				//% "Time-to-SoC 100%"
-				text: "Time-to-SoC 100%"
-				allowed: defaultAllowed && dataItem.seen
-				dataItem.uid: root.battery.serviceUid + "/TimeToSoC/100"
-				secondaryText: dataItem.isValid && dataItem.value != "" > 0 ? dataItem.value : "--"
+				secondaryText: Utils.secondsToString(dataItem.value)
 			}
 
 			ListRelayState {
@@ -303,6 +268,106 @@ Page {
 
 			ListAlarmState {
 				dataItem.uid: root.battery.serviceUid + "/Alarms/Alarm"
+			}
+
+			ListNavigationItem {
+				text: "dbus-serialbattery - General"
+				allowed: defaultAllowed && cvl.isValid || ccl.isValid || dcl.isValid
+				/*
+				allowed: defaultAllowed && {
+					// mr-manuel/dbus-serialbattery
+					productId.value === 0xBA77
+					// Dr-Gigavolt/dbus-aggregate-batteries
+					|| productId.value === 0xBA44
+				}
+				*/
+				onClicked: {
+					Global.pageManager.pushPage("/pages/settings/devicelist/battery/PageBatteryDbusSerialbattery.qml",
+							{ "title": text, "bindPrefix": root.battery.serviceUid })
+				}
+			}
+
+			ListNavigationItem {
+				text: "dbus-serialbattery - Cell Voltages"
+				allowed: defaultAllowed && cell3Voltage.isValid
+				onClicked: {
+					Global.pageManager.pushPage("/pages/settings/devicelist/battery/PageBatteryDbusSerialbatteryCellVoltages.qml",
+							{ "title": text, "bindPrefix": root.battery.serviceUid })
+				}
+
+				VeQuickItem {
+					id: cell3Voltage
+					uid: root.battery.serviceUid + "/Voltages/Cell3"
+				}
+			}
+
+			ListNavigationItem {
+				text: "dbus-serialbattery - Settings"
+				// show only for mr-manuel/dbus-serialbattery (productId registered at Victron)
+				allowed: defaultAllowed && productId.value === 0xBA77
+				onClicked: {
+					Global.pageManager.pushPage("/pages/settings/devicelist/battery/PageBatteryDbusSerialbatterySettings.qml",
+							{ "title": text, "bindPrefix": root.battery.serviceUid })
+				}
+			}
+
+			ListNavigationItem {
+				text: "dbus-serialbattery - Time to SoC"
+				allowed: defaultAllowed && timeToSoc0.seen ||
+						timeToSoc5.seen ||
+						timeToSoc10.seen ||
+						timeToSoc15.seen ||
+						timeToSoc20.seen ||
+						timeToSoc80.seen ||
+						timeToSoc85.seen ||
+						timeToSoc90.seen ||
+						timeToSoc95.seen ||
+						timeToSoc100.seen
+				onClicked: {
+					Global.pageManager.pushPage("/pages/settings/devicelist/battery/PageBatteryDbusSerialbatteryTimeToSoc.qml",
+							{ "title": text, "bindPrefix": root.battery.serviceUid })
+				}
+
+				VeQuickItem {
+					id: timeToSoc0
+					uid: root.battery.serviceUid + "/TimeToSoC/0"
+				}
+				VeQuickItem {
+					id: timeToSoc5
+					uid: root.battery.serviceUid + "/TimeToSoC/5"
+				}
+				VeQuickItem {
+					id: timeToSoc10
+					uid: root.battery.serviceUid + "/TimeToSoC/10"
+				}
+				VeQuickItem {
+					id: timeToSoc15
+					uid: root.battery.serviceUid + "/TimeToSoC/15"
+				}
+				VeQuickItem {
+					id: timeToSoc20
+					uid: root.battery.serviceUid + "/TimeToSoC/20"
+				}
+				VeQuickItem {
+					id: timeToSoc80
+					uid: root.battery.serviceUid + "/TimeToSoC/80"
+				}
+				VeQuickItem {
+					id: timeToSoc85
+					uid: root.battery.serviceUid + "/TimeToSoC/85"
+				}
+				VeQuickItem {
+					id: timeToSoc90
+					uid: root.battery.serviceUid + "/TimeToSoC/90"
+				}
+				VeQuickItem {
+					id: timeToSoc95
+					uid: root.battery.serviceUid + "/TimeToSoC/95"
+				}
+				VeQuickItem {
+					id: timeToSoc100
+					uid: root.battery.serviceUid + "/TimeToSoC/100"
+				}
 			}
 
 			ListNavigationItem {
@@ -317,20 +382,6 @@ Page {
 				BatteryDetails {
 					id: batteryDetails
 					bindPrefix: root.battery.serviceUid
-				}
-			}
-
-			ListNavigationItem {
-				text: "Cell Voltages"
-				allowed: defaultAllowed && cell3Voltage.isValid
-				onClicked: {
-					Global.pageManager.pushPage("/pages/settings/devicelist/battery/PageBatteryCellVoltages.qml",
-							{ "title": text, "bindPrefix": root.battery.serviceUid })
-				}
-
-				VeQuickItem {
-					id: cell3Voltage
-					uid: root.battery.serviceUid + "/Voltages/Cell3"
 				}
 			}